requirements.txt
Django==4.2.7
django-cors==4.3.1
djangorestframework==3.14.0
django-environ==0.11.2
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1